[Relations] [Benwei] 完成PIService对OneParent计算的实现，修改了AlleleValueDao的取值方式，提高所有的数值变量精度到double

+++ src/main/java/me/zeph/relations/model/PIValue.java
@@ -2,17 +2,17 @@ package me.zeph.relations.model;
 
 public class PIValue {
 
-	private float value;
+	private double value;
 
-	public PIValue(float value) {
+	public PIValue(double value) {
 		this.value = value;
 	}
 
-	public float getValue() {
+	public double getValue() {
 		return value;
 	}
 
-	public void setValue(float value) {
+	public void setValue(double value) {
 		this.value = value;
 	}
 }
